feat: refactor types, password-reset.tsx const vars
