+++ src/MyWorkID.Client/src/components/ui/icons/temporary-access-pass-icon.tsx
@@ -1,6 +1,4 @@
-import type { ComponentPropsWithoutRef } from "react";
-
-type TemporaryAccessPassIconProps = ComponentPropsWithoutRef<"svg">;
+import { SvgIconProps } from "../../../types";
 
 export const TemporaryAccessPassIcon = ({
   width = 20,
@@ -12,7 +10,7 @@ export const TemporaryAccessPassIcon = ({
   strokeLinecap = "round",
   strokeLinejoin = "round",
   ...svgProperties
-}: TemporaryAccessPassIconProps) => {
+}: SvgIconProps) => {
   return (
     <svg
       width={width}
